AI-Powered Personal Finance Advisor/finwise-ai/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
prophet==1.2.0
openai==1.14.3
python-dotenv==1.0.1
plaid-python==9.0.0
